Milestone 5: slice builtins — make!, append!, copy!, slice!, cap
Brings Go's slice-shaped builtins online so call sites read as Go
verbatim except for the `!` macro suffix:

    let xs = slice!([]int{1, 2, 3});
    let ys = make!([]int, 5);                // 5 zeros
    let buf = make!([]byte, 0, 1024);        // empty, cap 1024
    let xs = append!(xs, 4, 5, 6);           // variadic
    let n  = copy!(dst, src);                // returns int
    let c  = cap(&xs);                       // free fn over Cap trait

Macros are needed (over free functions) because Go's `make([]T, ...)`
and `[]T{...}` carry a *type argument* in their syntax, and `append`
is variadic — Rust functions can't model either without macros.

Files:
  - builtin.rs: cap() free function over a new Cap trait + the
    __make_size length-checker that panics on negative size (matches
    Go's runtime check).
  - builtin_macros.rs: slice!, make!, append!, copy! macros.
  - lib.rs: extern crate alloc (already there); plus a hidden
    __macro_alloc module re-exporting Vec / vec! so user binaries
    don't need their own `extern crate alloc;` to call the macros.
  - goslice.rs: Cap trait impl, DerefMut so copy! auto-borrows the
    destination, __from_vec / __into_vec hidden methods replacing the
    pub(crate) versions (macros need them to resolve at user-call-site
    paths).
  - gostring.rs: __from_vec rename. From<&'static str> for `string` so
    slice!([]string{"a", "b"}) works via .into().
  - types.rs: int = i64, uint = u64. We pin to amd64 width (Go's
    `int` is 64-bit on amd64) so `From<i32>` works in core for
    .into()-based macro arms — `From<i32>` for `isize` is not
    implemented in core. Documented in the type comment.
  - convert.rs / unicode/utf8/mod.rs: callers updated to the renamed
    __from_vec / __into_vec.
  - examples/slices.rs: 12-check smoke test exercising literal,
    make, append, copy, range over slice, bytes round-trip.

Release sizes: hello 1 KiB, alloc_smoke 14 KiB, strings 19 KiB,
slices 15 KiB. All examples statically linked, no libc symbols.

+++ src/types.rs
@@ -4,16 +4,21 @@
 //   ─────────────   ──────────────
 //   byte = uint8    byte = u8
 //   rune = int32    rune = i32
-//   int             int  = isize    ← platform-sized signed (Go's `int`)
-//   uint            uint = usize
+//   int             int  = i64       ← matches Go's `int` width on amd64
+//   uint            uint = u64
 //
-// `int` matches Go's int width on amd64 (64-bit). `rune` is signed
-// just like Go's `type rune = int32`, even though valid code points
-// are non-negative — this keeps literal arithmetic well-behaved.
+// We pin `int = i64` (rather than `isize`) so `From<i32>` and friends
+// are implemented in core, which makes `.into()` Just Work in the
+// `slice!`/`append!` macros for integer literals. v1 targets 64-bit
+// Linux only, so this matches Go's int width exactly. (On a true
+// 32-bit port, this alias would change to i32; v1 defers that.)
+//
+// `rune` is signed like Go's `type rune = int32`, even though valid
+// code points are non-negative — keeps literal arithmetic well-behaved.
 
 #![allow(non_camel_case_types)]
 
 pub type byte = u8;
 pub type rune = i32;
-pub type int = isize;
-pub type uint = usize;
+pub type int = i64;
+pub type uint = u64;
